Added support for mounting another inode as root
The super block stores a default root which sets the effective root
unless it is overwritten by the 'root=' mount option. Although the
directory entry for '..' might not point to itself as is expected of the
root directory, Linux seems to overwrite this anyway so I left the
directory handling code as is.